docusaurus_app/backend/src/services/db_service.py
import os
from sqlalchemy import create_engine, text
from sqlalchemy.orm import sessionmaker
from typing import Optional, Any

class DBService:
    def __init__(self):
        self.neon_db_url = os.getenv("NEON_DATABASE_URL")
        if not self.neon_db_url:
            raise ValueError("NEON_DATABASE_URL environment variable not set.")
        
        self.engine = create_engine(self.neon_db_url)
        self.SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=self.engine)
        print("Neon Postgres DBService initialized.")

    def get_session(self):
        db = self.SessionLocal()
        try:
            yield db
        finally:
            db.close()

    def execute_query(self, query_str: str, params: Optional[Dict[str, Any]] = None):
        with self.engine.connect() as connection:
            result = connection.execute(text(query_str), params)
            connection.commit()
            return result
            
    # Placeholder methods for user and preference interaction
    def create_user(self, user_data: Dict[str, Any]) -> Dict[str, Any]:
        print(f"Placeholder: Creating user with data: {user_data}")
        # In a real implementation, this would insert into a users table
        return {**user_data, "user_id": "mock_user_id"}

    def get_user(self, user_id: str) -> Optional[Dict[str, Any]]:
        print(f"Placeholder: Getting user with ID: {user_id}")
        # In a real implementation, this would query a users table
        if user_id == "mock_user_id":
            return {"user_id": user_id, "name": "Mock User", "email": "mock@example.com"}
        return None

    def update_user_preference(self, user_id: str, preference_data: Dict[str, Any]) -> Dict[str, Any]:
        print(f"Placeholder: Updating preferences for user {user_id} with: {preference_data}")
        # In a real implementation, this would update/insert into a user_preferences table
        return {**preference_data, "user_id": user_id, "preference_id": "mock_pref_id"}

    def get_user_preferences(self, user_id: str) -> Dict[str, Any]:
        print(f"Placeholder: Getting preferences for user {user_id}")
        # In a real implementation, this would query a user_preferences table
        return {"preferred_language": "en", "experience_level": "intermediate"}

